chore: 完成多模块迭代更新与修复
本次提交包含大量功能迭代与bug修复：
1. 订单模块：新增退款失败重试任务、调整订单列表页大小、修复支付回调路由、添加退款记录重试查询方法
2. 会员模块：调整会员列表默认分页、优化会员过期处理、更新会员配额与成长日志模型
3. 消息模块：新增未读数缓存刷新任务、修复公告状态码与发送逻辑、优化通知偏好设置
4. 反馈模块：调整权限路由、修复低分告警队列与优先级、优化评价创建事务与空值处理
5. 预测/评估模块：移除重复防重中间件、优化任务状态返回格式
6. 套餐优惠券模块：修复优惠券删除校验、调整权限标识、优化表单校验规则
7. 数据模型：新增退款重试次数字段、修复CouponStatsVO与OrderStatsVO的使用率/退款率类型
8. 其他：修复导入导出错误码、优化用户标签查询、修复任务创建参数、清理过期消息条件

+++ dehaze-go/internal/model/vo/order.go
@@ -76,7 +76,7 @@ type OrderStatsVO struct {
 	TotalOrders           int64                      `json:"totalOrders"`
 	TotalRevenue          int64                      `json:"totalRevenue"`
 	TotalRefund           int64                      `json:"totalRefund"`
-	RefundRate            int64                      `json:"refundRate"`
+	RefundRate            float64                    `json:"refundRate"`
 	StatusDistribution    map[string]int64           `json:"statusDistribution"`
 	PayMethodDistribution map[string]int64           `json:"payMethodDistribution"`
 	PackageDistribution   []OrderPackageStatItem     `json:"packageDistribution"`
